pyproject.toml
[project]
name = "swagger-py-annotation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
sspa-export = "ssap.tools.cli:main"
